Sequencer: Add middle click copy key from maya feature.
Middle click on a key copies that key and initiates dragging of the new key. Middle click on an empty key area will create a key and set the value at the curve's value at that time. Key groupings are supported as well (only channels that are keyed are copied).

[CL 2689610 by Max Chen in Main branch]