Added new blueprint alignment algorithm: straighten connections
This allows straightening of connections between all selected nodes in a graph, as well as specific connections between pins (by right clicking on the pin).
Existing alignment routines now align to the clicked-on node, where applicable, and use an average where this is not available.
Added icons for all blueprint alignment commands.

[CL 2690063 by Andrew Rodham in Main branch]