attach weldness of being crash
Some sort of change in collision status, attachment, welding configuration was causing a recreate of physx objects.  because of the way things were configured, it was attempting to initialize a body twice into physx.   the first time due to an unweld would instantiate the "root' which turned out to be self, and then again later it detected that it was actor was already created but not properly dealt with.  now we remove things from the to-be-created-and-added-to-scene list if they happen to magically appear during the welding steps.

http://crashreporter/Buggs/Show/70202
#jira UE-20075

[CL 2692798 by Stan Melax in Main branch]