Suspending the undo buffer during duplication of a Blueprint's CDO during reinstancing until the CDO is fully resolved.
During duplication of the CDO, the duplicated CDO's class still does not point to the CDO when it is placed in the transaction buffer. Any native sub object components that are placed in the transaction buffer are serialized incorrectly. These are fixed up during a compile which leads to a different path of serialization during undo and a mismatch buffer size. Delaying CDO placement into the undo buffer until it is finished resolves this.

#jira UE-20913 - Undo after deletion in BP editor causes crash

#codereview Maciej.Mroz, Phillip.Kavan

[CL 2697003 by Michael Schoell in Main branch]