[AUTOMERGE]
Added new property to all EnvQueryTests: TestComment.  This is a simple string used to optionally label the purpose for a test.

For now, it only shows up in the details view, but eventually we will surface this comment for users in a better UI manner so it can be easily seen when viewing the EQS data asset.

#ReleaseNoteNeeded

--------
Integrated using branch Ue4-To-UE4-Fortnite-Simple (reversed) of change#2699000 by Daniel.Broder on 2015/09/21 10:56:38.

[CL 2699005 by Daniel Broder in Main branch]