Improvements to Find-in-Blueprints for variable nodes.
Properties in nodes can be marked as Blueprint searchable. Content will be gathered by FiB system for use in unloaded Blueprints on resave.

"Find References" on variables in the MyBlueprint list or on variable nodes is now more accurate.

Removed FEdGraphSchemaAction_K2AddCallOnVariable and FEdGraphSchemaAction_K2AddDocumentation, both unused in the editor and are a part of the old Blueprint menu system.

FMemberReference::SetSelfMember and FMemberReference::SetExternalMember support accepting a FGuid for the reference.

Variable nodes will assign their variable's Guid to their VariableReference
Added UK2Node::GetFindReferenceSearchString to find the search query for a node to find references to it, currently returns the node title for all nodes but variables (where it tries to do an advanced search to limit to only actual references to the same variable).

Can mark UProperties with metadata that informs the Find-in-Blueprint system to gather the property for searching. Currently FVariableReference in K2Node_Variable is the only thing gathered and only nodes are yet supported by this feature. Will expand to all objects contained within UBlueprint in the future.

When using Find-in-Blueprints, can use category names as functions to filter into child tags and values.

All Find-in-Blueprint data is now stored in an Asset Registry tag titled "FiBData" instead of "FiB", this is to support the new version feature.

#jira UE-21164 - Find-in-Blueprints needs an easier way of marking properties as searchable.

[CL 2702574 by Michael Schoell in Main branch]