Added new blueprint alignment algorithm: straighten connections
This allows straightening of connections between all selected nodes in a graph, as well as specific connections between pins (by right clicking on the pin).
Existing alignment routines now align to the clicked-on node, where applicable, and use an average where this is not available.
Added icons for all blueprint alignment commands.

#jira UE-21428

[Merged from 2690063 using UE4-To-UE4-Releases-4.10]

[CL 2703956 by Andrew Rodham in 4.10 branch]