Made the Localise automation script more data driven, and removed some hard-coded assumptions
The Localise automation script processing was failing as we have acquired the "zh-Hans-CN" culture in our OneSky project, but the UE4 commandlet doesn't generate that culture, only "zh-CN". This was causing the script to fail as it couldn't find the data for that culture in Perforce.

This change has the Localise automation script read the same config file that the UE4 commandlet will use, and then use that config file data to work out what cultures it will export/upload to OneSky.

This change should also make the script generic enough for it to be used for other projects (once we can pass in the arguments on the command line).

[CL 2704649 by Jamie Dale in Main branch]