[AUTOMERGE]
Attempt to spawn a PlayerState for the replay spectator controller during client replay recording, by falling back to the default game mode object if the authority game mode can't be found (only if GetNetMode() != NM_Client).
Expose the ability to get an actor's FNetworkGUID, and the actor that corresponds to an FNetworkGUID, if available. Allows games to correlate actors across multiple local worlds if there is a client-server relationship between them.

--------
Integrated using branch UE4-Orion-To-UE4 of change#2705235 by Ryan.Gerleve on 2015/09/24 17:12:18.

[CL 2705290 by Ryan Gerleve in Main branch]