#codereview Jamie.Dale, Justin.Sargent
Largely rewritten Editor Region & Language Settings page. Far less error prone compared to the old implementation. I'd like to do a lot more to make this not so lame in the future - move the game's native culture to something not editor-specific and leverage it as the default for new game targets in the localization dashboard.

[CL 2706823 by Saul Abreu in Main branch]