Fixed FTextSnapshot reporting all transient text instances as dirty
FTextSnapshot was changed to use a TWeakPtr, however this meant that for text that was generated each frame, the FTextSnapshot would always report that it was out-of-date (even if the generated text was the same as the previous frame) as it no longer had a valid reference point to compare against.

FTextSnapshot has been changed to once again use a TSharedPtr, which will keep the transient instance around long enough for it to be compared with the text from the next frame.

[CL 2707568 by Jamie Dale in Main branch]