Fixed some FText instances not rebuilding correctly when the culture is changed
FTextHistory has an IsOutOfDate function that FTextHistory_Base was overriding to prevent any text rebuilding from happening. This meant that the recent changes to FText::Rebuild had stopped any FTextHistory_Base instances from ever reporting themselves as dirty, and so they didn't update when the culture was changed.

This change separates the concerns of a history being "out of date" (ie, not at the head revision of the localisation manager), and "allowing a text rebuild" (now handled by FTextHistory::CanRebuildText). This fixes the issue where some texts would fail to update when the culture was changed.

#jira UE-21549

[CL 2709156 by Jamie Dale in Main branch]