Sequencer marquee selection improvements
Marquee selection now intelligently selects keys and sections depending on the context. If a section is within the marquee it is eligible for selection, unless there are keys from that section also in the marquee that would change selection state as a result of the operation.

Selections are also now previewed in real-time

[CL 2714290 by Andrew Rodham in Main branch]