#jira UE-22327
Fixed up 4.10 TP_SideScroller (C++) template by removing the UPROPERTY tag from the offending variable and then resaving the map.

It is likely that there is still an underlying problem here. I'm not clear that what was done should cause an error message, given that it worked fine. So, either something is broken in an ignorable way, or the error message is coming up when it shouldn't.

[CL 2748074 by Richard Hinckley in Main branch]