UE4.11 - D3D12 integration & mac compile fix
Integration from changelist 2800294
Update how command lists are batched and submitted to the GPU. I added a console variable so it can be controlled at runtime. The aggressive batching mode only have 2 ExecuteCommandList() calls/frame, similiar to single threaded mode. One danger of aggressive batching is that you can starve the GPU so developers will need to pay close attention to this and tune the batching for their application.
#jira UE-24967

[CL 2817098 by Rolando Caloca in 4.11 branch]