integrated from Dev-General:
new feature: detect bad video card driver and warn user, currently 2 entries (could/should be made data driven). Currently implement in UE4 but if needed could be made game specific or moved into the Portal (not ideal as it doesn't detect all cases)
-----------
Send GPU RHI adpater over HardwareSurvey and GameFeedback
-----------
In case of Windows driver bugs we might want to blacklist a driver or show a UI to warn the user.
This code adds the capability to get the driver version and shows it in the log
(early code, it might fail to find the right driver, then we cannot rely on the data, string should be empty):
Example log printout:

0. 'AMD Radeon R9 200 / HD 7900 Series' (Feature Level 11_0)
    3051MB of dedicated video memory, 0MB of dedicated system memory,
    and 3840MB of shared system memory, 2 output[s]
1. 'Microsoft Basic Render Driver' (Feature Level 11_0)
    0MB of dedicated video memory, 0MB of dedicated system memory,
    and 256MB of shared system memory, 0 output[s]

Chosen D3D11 Adapter: 0

   Adapter Name: AMD Radeon R9 200 / HD 7900 Series
 Driver Version: 15.7.1 (internal 15.200.1062.1004)
    Driver Date: 8-3-2015

 Next: test on various cards, add UI to appear if certain driver version is detected
#rb:Olaf.Piesche
#jira:OR-12462 and others

[CL 2828755 by Martin Mittring in 4.11 branch]