Crash fix when undo/redo'ing the paste of a event node.
Fixed issues where when pasting nodes the entire structural recompile of the BP was being captured in the transaction.

Event nodes that are replaced on paste with custom event nodes will now be marked as pending kill.

Getting a node's graph will no longer ensure if the node is marked as pending kill.

#jira UE-26004 - Crash when redoing a paste of a copied event node

[CL 2845806 by Michael Schoell in 4.11 branch]