Merging from //UE4/Dev-Editor CL 2829387
Fixed crash when using the eye dropper on color pickers in material editor graph nodes.
#jira UE-24871 - Crash upon selecting the eyedropper in the color picker within the material editor

Previously the callback to the material editor (when the eyedropper color was changed) was happening in an active tick, but this was causing the cached material render proxy, stored by FPreviewElement::BeginRenderingCanvas, to become invalid, as the change of preview material would be taking place during the Slate tick/render phase.  Moving this callback to PreTick fixes this, so that the Preview Viewport for the material graph node is able to get an OnPaint (and cache the latest preview material) prior to performing the actual render.

[CL 2870009 by Richard TalbotWatkin in 4.11 branch]