merging from //UE4/Dev-Platform to //UE4/Release-4.11
--- original commit CL: #2913300 message ---
corrected VS 2015 websocket lib to look at the right offset

it is currently a high risk change to just update the libwebsocket wholesale for  release-4.11.

this change is the most minimum invasive change with a lot of deep analysis (details will be put in jira: # UEPLAT-1221).

this fix will be pushed up to main and release-4.11

#jira UE-22166 - HTML5 Cook on the fly will launch and then close browser
#jira UE-22513 - HTTP Network File System crashes randomly.
#jira UE-28003 - Fail to QuickLaunch HTML5 through UnrealFrontEnd

[CL 2913310 by Nick Shin in 4.11 branch]