Duplicating CL#2838289 from //UE4/Dev-Core
#JIRA UE-25748

When loading an object with duplicated lazy ptr GUID, don't create a new GUID but invalidate the GUID to avoid undeterministic cooked builds.

Previous behaviour created a new GUID for the object which didn't really matter because there were not lazy ptrs in the engine that would use it anyway.
So setting the GUID to zero has exactly the same effect as far as lazy pointers are concerned with the exception that we're not producing new GUIDS.

[CL 2919898 by Bob Tellez in 4.11 branch]