Fix fast path corner case issues
The anim BP compiler was serializing out the size of a value to be copied at runtime. This varies on platforms other than the one that is doing cooking. We now re-calculate the size on load.
Also, boolean properties operated on by the fast path could fail to get/set their values correctly if they were bitfield properties rather than standard booleans. We now use Set/GetPropertyValue_InContainer to operate on boolean properties.
Also removed untested code path for source array properties (no assets we can find use this). This now falls back to the BP path.

Note: Tweaked from the original version submitted in CL 2934734 so it is applicable as a hotfix to 4.11.

#rb Martin.Wilson
#jira UE-29116 - Animation CopyRecords need to calculate size at runtime
#lockdown Ben.Marsh

[CL 2940647 by Thomas Sarkanen in 4.11 branch]