Fix layered bone per blend accessed from a struct in the fast-path
Made sure that the fallback event is always built (logic was still split so if PatchFunctionNamesAndCopyRecordsInto aborted because of some unhandled case if might not have an event to call).
Covered struct source->array dest case.
Indicator icon is now built from the copy record itself, ensuring it is accurate to actual runtime data.

#jira UE-29389 - Fast-Path: Layered Blend per Bone node failing to grab updated values from struct.
#rb Martin.Wilson

[CL 2949739 by Thomas Sarkanen in 4.12 branch]