Fixed -windowed using the wrong window size
This was being clamped to the client size of the window, rather than the screen size of the window. This was also preventing the resolution being increased from the initial size.

We now consider the max size of a window to be its client size, and we use the maximum of the display area or requested resolution as the maximum window size (as we sometimes create 4K windows to capture).

#jira UE-30112

[CL 2966056 by Jamie Dale in 4.12 branch]