Fixed an issue where triggering a modal window could cause the modifier key cache to be wrong
The modal window can cause us to consume the key-up notification, which for some reason also wasn't being updated when the window activation was changed. We now update the modifier key states when we consume a key input message due to a modal window.

This also removes the logic duplication between PressedModifierKeys and CachedModifierKeyState. We now have one fixed size array (ModifierKeyState) which holds whether or not the modifier keys are pressed (or toggled in the case of caps lock). This information is then used to build the FModifierKeysState used by Slate.

#jira UE-28172
#rb Marc.Audy

[CL 2967885 by Jamie Dale in 4.12 branch]