Fixed recorded animation trasform issues & misc. other recording bugs
The problem was that some assets in QAGame had rotated root bones. Because we use root motion mask out recorded root transforms, this ended up giving an incorrect transform to the actor when we extracted the root motion post-record. To fix this, we store the inverse initial root transform and then apply it to the root and its children to fixup the initial offset.
Handled slave components & components without SkelMeshes a more throughly (Fortnite recording now works).
Handled visibility of added components (adds a key at the start to handle pre-roll extrapolation correctly).

#jira UE-30295 - Sequence recording generates incorrectly transformed components

[CL 2968779 by Thomas Sarkanen in 4.12 branch]