Migrating fix from DevRendering:
Only allocate reflection capture cubemaps if we are actually doing a reflection capture. The old approach always allocated based on CVar and worked with a warning on D3D since if SRC was larger than the DST, the extra SRC area was dropped. New approach only allocates when necessary and is correctly sized everytime. Also hardened access to CVar and what users are allowed to set.

#lockdown cristina.riveron
#jira UE-30985

[CL 3000701 by Zabir Hoque in 4.12 branch]