#jira UEDOC-3005
Added the following Metadata in order to comply with the latest version of the Style & Standards Guide:
- Parent
- Order
- Type

Updated engine version to reflect latest stable release.
Added Table of Contents in order to comply with the latest version of the Style & Standards Guide.
Various minor grammatical and stylistic revisions per my Peer Review comments.

#UE4docs

[CL 3056059 by Robert Gervais in Main branch]