#jira UEDOC-1978
Added CPP FPS Button Object on the Programming Tutorials Landing Page.
The button object is currently commented out because it is still in Editorial review; however, the block can be uncommented when the CPP FPS Tutorial is ready for publication.
Updated the CPP FPS Landing Page Description to be a bit more descriptive.

#UE4docs

[CL 3064949 by Robert Gervais in Dev-VREditor branch]