Slate - Reverting the SMenuAnchor to a previous version, there was no reason afterall to need to use the last painted window as the host for menus, ended up solving it a lower level by properly supporting the deferral groups on the SVirtualWindow under different conditions.
Slate - The hit test grid now properly records the hit test path for the invalidation box, so that when input is recieved, a widget path containing only one instance of the invalidation box is created, premitting things like mouse capture to properly work.

UMG - Further refinements and improvements to the Widget Interaction Component.  This completes the documentation and a fixes several bugs with it that were found after the integration to main occured.

#rb none
#jira UE-33845

[CL 3068204 by Nick Darnell in 4.13 branch]