Sequencer: Fixed sequencer adding spawnables into recorded worlds twice
The issue here is that sequencer was re-evaluating itself with the current play world as the context, which also happened to be the world that it was trying to record. As a result, it ended up with the same sequence being played twice in the recording world.

Added the ability to specify a playback context attribute for sequencer to use, and this no longer allows sequencer to play back in any worlds that are being recorded.

#jira UE-31422
#rb Max.Chen

[CL 3077001 by Andrew Rodham in 4.13 branch]