Fixed a potential race-condition in FTextRenderComponentMIDCache, and updated it to detect "stale" MIDs
FMIDData was shared between the game and render threads, but used non-thread-safe shared pointers.

This also marks MIDs as "stale" if the number of MIDs no longer matches the number of pages in the font (which may happen if the font is edited). These "stale" MIDs are kept as a weak pointer in a separate array so that we can still keep the MID object alive as long as something is still using it (as it may still be used by a FTextRenderSceneProxy for a short while). This array of weak pointers is purged of unreferenced instances during the normal cache purge cycle.

#rb Matt.Kuhlenscmidt
#jira UE-37519
#lockdown Nick.Penwarden

[CL 3170815 by Jamie Dale in 4.14 branch]