#jira UE-38072 #rb Nick.Whiting #lockdown nick.penwarden
We need to add a hook that can be called after native present has finished for SteamVR.

PostPresentHandoff should be called when using the interleaved compositor immediately after we've submitted our eye buffers and called present for the mirror window. This unblocks the compositor process so it can do it's re-projection work. Otherwise it will block until we call WaitGetPoses which is a ways into the next frame.

[CL 3181939 by Ryan Vance in 4.14 branch]