#jira UE-UE-38161 #rb Nick.Whiting #lockdown nick.penwarden
Adding ISceneViewExtension::UsePostInitView which will be used to enable/disable the usage of PostRenderViewFamily_RenderThread and PostRenderView_RenderThread which was added earlier. We need to enable this behavior based on the hmd's compositor behavior, so a simple cvar wont work.
I missed the PostPresent implementation for steamvr in the earlier check in.

[CL 3184418 by Ryan Vance in 4.14 branch]