Copy of CL 3162062
Replaced FWindowsCursor hack for warping the mouse cursor to the center of the viewport with a better fix for the original problem (users being able to resize the window while the cursor is hidden and the mouse controls the camera). This change removes round window corners in borderless window mode and disables window resizing when the cursor is hidden.

#rb Matt.Kuhlenschmidt
#jira UE-38721
#lockdown cristina.riveron

[CL 3202043 by Michael Trepka in 4.14 branch]