Fixed asset attachment issues in Skeleton Tree
Assets were being attached uniquely, so only one asset could be attached to a bone/socket. However the calling code didnt know that the unique attachment function just gave up, so the item just got added to the bottom of the tree.
The attachment filter was not set correctly to allow for bone attatchments, so only sockets could be attached to.
The attach parent name was not initialized, so assets could not be deleted one at a time.

#jira UE-40040 - With multiple Preview assets on one bone, only one appears in Skeleton Tree
#jira UE-40041 - Preview assets appear at the bottom of the skeleton tree
#lockdown Cristina.Riveron
#rb Keith.Judge

[CL 3246005 by Thomas Sarkanen in 4.15 branch]