Duplicated from //ue4/Release-4.14
CL 3238182
Disable timestamp queries on pre-Maxwell nvidia hardware. Local testing suggests that this is the major cause of instability in the UE4.14 release.

It's possible that we could be more targeted by only excluding Fermi and older hardware, but identifying fermi hardware by device ID is difficult in practice, since the range overlaps with Kepler.

#lockdown nick.penwarden
#jira UE-38818
#rb marcus.wassmer

[CL 3250830 by Ben Woodhouse in 4.15 branch]