Merging 3232641 & 3236788 & 3233854 & 3249742 from Dev-Rendering:
3232641:
- Eliminate redundant state changes in MetalRHI in the state cache.
- Add a new debug level for setting buffers to nil prior to calls to set*Bytes so that the tool doesn't display incorrect data.
- Make testing for validation & statistics features use the same EMetalFeatures API as everything else for consistency.
- Cache the fallback depth-stencil texture in the state cache and ignore it for determining whether a pass can restart - if we are using this texture its contents are worthless anyway.
#rb rolando.caloca

3236788:
Fix 10.11.6 support (aka -nometalv2): the stencil view workaround necessitates a mid-render blit and the way things were setup resulted in the HasValidRenderTargets assert firing. Refactored the code to separate the concept or valid render-states in the cache from active render-states in the render-pass. Now it works as intended and will be needed for 4.15.
#rb michael.trepka

3233854:
More information about texture type validation errors in Metal.

3249742:
Fix missing GPU particles on Mac.
Pointers getting reused is causing the blendstate equality operator to fail.
Simple workaround until we have time for a proper fix.
#rb Michael.Trepka
#lockdown gil.gribb

#jira UE-40200

[CL 3253807 by Mark Satterthwaite in 4.15 branch]