Added extra context to crash analytics and crash reports in the Editor. [UE-41306] - Add context to crash analytics and crash reports (with Editor user activity)
When a Slate tab becomes active or foregrounded, we take its LayoutIdentity, Label and Content Widget Type and generate a string from all three. This gives context for what the user was doing. The string is set as the UserActivity and passed to MTBF analytics, crash analytics and crash reporter.

Also added filter to the UserActivity tracking system that defaults to Game mode to preserve previous behavior. Editor now switches it to Editor mode and starts setting the activity in this mode.

#jira UE-41306
#rb none
#codereview Matt.Kuhlenschmidt

[CL 3278667 by Chris Wood in 4.15 branch]