Mirroring CL 3393509 from Dev-Framework
Removing hack'ish ResetLoaders() call that was causing undesired side-effects (resetting of a loaded package that other objects were relying on). This was originally intended to release file handles so separate editor processes could make updates and save the file (from CL 1712376). Using ResetLoaders() for this is bad though, as it has too many side effects. Instead we have to wait for GC to run. This also makes sure that GC should run as intended as the CookOnTheFly sever is idling.

#jira UE-37284
#rb Daniel.Lamb
#lockdown cristina.riveron

[CL 3398112 by Mike Beach in 4.16 branch]