Fixed a crash when reloading a curve asset that was open in an embedded curve editor
Curves are edited via an interface as not all of them are assets. This change adds a RepointCurveOwner function to the interface which can be used by the curve editor to repoint its curve owner during a package reload. This is needed as the assets use multiple inheritance, meaning the UObject and FCurveOwnerInterface pointers are offset, so we can't perform a simple pointer repoint directly in the curve editor.

#jira UE-46699
#rb Matt.Kuhlenschmidt

[CL 3521976 by Jamie Dale in 4.17 branch]