#jira UE-50505 #rb none
The depth target swap chain only creates a single SRV for stencil. We should probably move the stencil SRV into the depth texture that the swap chain aliases so this is handled like the depth SRV.

This is a smaller change to work around the issue for now. We make a new stencil SRV every frame to ensure it matches the active depth target when rendering in stereo and the HMD is allocating the depth target.

[CL 3676053 by Ryan Vance in 4.18 branch]