glTF: base64 decoding of data buffers and images
Data can be encoded inside the glTF JSON as a Base64 data URI. This CL addresses a known limitation of initial commit CL 3793018. I added this after because base64 is expected to be uncommon in the wild. Authoring software will typically use external BIN files (.gltf) or a BIN chunk (.glb) to store data.

#jira UE-50695
#rb none

[CL 3838909 by Mike Erwin in 4.19 branch]