Enable picking via Enter for details panel asset pickers
The previous (4.18) behavior was to only allow selection of the previous/next item in the list with arrow keys. A fix (CL 3783114) for pickers with many assets broke this. This change now allows for selection of any item by navigating with arrow keys and pressing Enter, as suggested in the Jira.

#jira UE-53440 - Unable to select assets within modal Asset Selection dropdowns after navigating with keyboard entry
#rb Jurre.deBaare

[CL 3843657 by Thomas Sarkanen in 4.19 branch]