Reduce editor startup times by ~15s on Windows.
Platform loading code recursively scans every module for dependent DLL modules to load first. Change to make it early-out as soon as it encounters a module which is already in memory (via a call to GetModuleHandle() from ResolveMissingLibraryImportsRecursive). Also use a TSet<> to store set of visited modules rather than an Array.

Now spends <0.1s total in this function on editor startup.

(Change looks larger than it is due to moving functions out of WindowsPlatformProcess.h to avoid introducing TSet dependency into this header).

#rb none
#jira UE-55642

[CL 3915240 by Ben Marsh in 4.19 branch]